+++ ride/profile.py
@@ -5,7 +5,6 @@ from typing import Any, Dict, Tuple, overload
 
 import numpy as np
 import torch
-from torch.autograd.profiler import format_time
 from tqdm import tqdm
 
 from ride.utils.gpus import parse_gpus, parse_num_gpus
@@ -133,3 +132,14 @@ class ProfileableDataset:
     @abstractmethod
     def profile(self) -> Dict[str, Any]:
         ...  # pragma: no cover
+
+
+def format_time(time_us):
+    """Defines how to format time in FunctionEvent"""
+    US_IN_SECOND = 1000.0 * 1000.0
+    US_IN_MS = 1000.0
+    if time_us >= US_IN_SECOND:
+        return "{:.3f}s".format(time_us / US_IN_SECOND)
+    if time_us >= US_IN_MS:
+        return "{:.3f}ms".format(time_us / US_IN_MS)
+    return "{:.3f}us".format(time_us)