+++ ride/utils/__init__.py
@@ -1,1 +1,1 @@
-from .utils import attributedict, some  # noqa: F401
+from .utils import attributedict, name, some  # noqa: F401